Added sub-directories model, controller and repository. Also further organise springboot backend files
