+++ backend/src/main/java/Controller/TaskController.java
@@ -1,5 +1,7 @@
-package TaskManager;
+package Controller;
 
+import Model.Task;
+import Repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -24,7 +26,7 @@ public class TaskController {
 
     // Creates a task with fields: title, description (optional), status and due date
     @PostMapping(path = "/add") // Map ONLY POST Requests
-    public @ResponseBody String addNewTask(@RequestParam String title, @RequestParam String description, @RequestParam String status, @RequestParam LocalDateTime dueDate) {
+    public @ResponseBody String addNewTask(@RequestParam String title, @RequestParam(required = false) String description, @RequestParam String status, @RequestParam LocalDateTime dueDate) {
         // @ResponseBody means the returned String is the response, not a view name
         // @RequestParam means it is a parameter from the GET or POST request
