+++ backend/src/main/java/Controller/TaskManagerApplication.java
@@ -1,12 +1,12 @@
-package TaskManager;
+package Controller;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
 @RestController
-public class DemoApplication {
+public class TaskManagerApplication {
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(TaskManagerApplication.class, args);
 	}
 }
