+++ backend/src/test/java/com/Shafi/TaskManager/TaskManagerApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class TaskManagerApplicationTests {
 
 	@Test
 	void contextLoads() {
